Decode and chunkprefill use the unified PyTorch API: flash_attn_with_kvcache

+++ src/sycl/kernels/flash_attention/xe_tile_scheduler.hpp
@@ -56,9 +56,9 @@ struct XeFHMAIndividualTileScheduler {
     using namespace cute;
 
     dim3 grid(
-        size(ceil_div(shape.head_size_vo, get<1>(tile_shape))),  // V
-        size(ceil_div(shape.seq_len_qo, get<0>(tile_shape))),    // Q
-        size(shape.batch * shape.num_heads_q));                  // (h,b) -- split later
+        size(ceil_div(shape.head_size_vo, get<1>(tile_shape))),     // V
+        size(ceil_div((int)shape.seq_len_qo, get<0>(tile_shape))),  // Q
+        size(shape.batch * shape.num_heads_q));                     // (h,b) -- split later
     return Params{grid, {shape.num_heads_q}};
   }
